requirements.txt
flask==3.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
werkzeug==3.1.3
apscheduler==3.10.4
requests==2.32.3
gunicorn==23.0.0
